Fix comparison between int and uint

+++ src/tokens_select.cpp
@@ -192,9 +192,9 @@ List qatd_cpp_tokens_select(const List &texts_,
     SetNgrams set_words;
     std::vector<std::size_t> spans = register_ngrams(words_, set_words);
     
-    if (pos_from_.size() != texts.size())
+    if (pos_from_.size() != (int)texts.size())
         throw std::range_error("Invalid pos_from");
-    if (pos_to_.size() != texts.size())
+    if (pos_to_.size() != (int)texts.size())
         throw std::range_error("Invalid pos_to");
     Positions pos(texts.size());
     for (size_t g = 0; g < texts.size(); g++) {
